Refactor: cts,python: Move base test classes into their own file.
Note that ValgrindTest could potentially also be considered a base test
class, but it's only used by one thing.  It might move into its own file
in the future.
