+++ python/pacemaker/_cts/tests/Makefile.am
@@ -1,0 +1,15 @@
+#
+# Copyright 2023 the Pacemaker project contributors
+#
+# The version control history for this file may have further details.
+#
+# This source code is licensed under the GNU General Public License version 2
+# or later (GPLv2+) WITHOUT ANY WARRANTY.
+#
+
+MAINTAINERCLEANFILES    = Makefile.in
+
+pkgpythondir = $(pythondir)/$(PACKAGE)/_cts/tests
+
+pkgpython_PYTHON = __init__.py \
+				   base.py
