+++ python/pacemaker/_cts/tests/__init__.py
@@ -1,0 +1,6 @@
+"""
+Test classes for the `pacemaker._cts` package.
+"""
+
+__copyright__ = "Copyright 2023 the Pacemaker project contributors"
+__license__ = "GNU Lesser General Public License version 2.1 or later (LGPLv2.1+)"
