+++ python/setup.py.in
@@ -16,5 +16,5 @@ setup(name='pacemaker',
       license='LGPLv2.1+',
       url='https://clusterlabs.org/pacemaker/',
       description='Python libraries for Pacemaker',
-      packages=['pacemaker', 'pacemaker._cts'],
+      packages=['pacemaker', 'pacemaker._cts', 'pacemaker._cts.tests'],
      )
